+++ conveyor/src/main/java/com/aryunin/conveyor/dto/EmploymentDTO.java
@@ -1,8 +1,8 @@
 package com.aryunin.conveyor.dto;
 
-import com.aryunin.conveyor.entity.enums.EmploymentStatus;
+import com.aryunin.conveyor.dto.enums.EmploymentStatus;
 import com.aryunin.conveyor.util.MoneySerializer;
-import com.aryunin.conveyor.entity.enums.Position;
+import com.aryunin.conveyor.dto.enums.Position;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
