+++ conveyor/src/main/java/com/aryunin/conveyor/dto/enums/Gender.java
@@ -1,4 +1,4 @@
-package com.aryunin.conveyor.entity.enums;
+package com.aryunin.conveyor.dto.enums;
 
 public enum Gender {
     MALE,
